🛠 update group members' data model
When a group owner's account is deleted, another member needs to be assigned as a new owner, which is when their photoUrl is necessary to update ownerPhotoUrl. Though the way of fetching group members' data also needs to be improved for a more efficient fetching, let's just leave it for now, as it is not must-have.